ALBS-5: Added ability to sign packages
This commit brings functionality to sign packages. This includes
database structure, tasks scheduling and changing statuses.
