+++ alws/alembic/versions/e88b665182b9_add_packages_signing.py
@@ -1,0 +1,53 @@
+"""Add packages signing
+
+Revision ID: e88b665182b9
+Revises: 585ab06efb32
+Create Date: 2021-12-05 22:22:57.128328
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = 'e88b665182b9'
+down_revision = 'a8007f32ebc7'
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('sign_keys',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('name', sa.Text(), nullable=True),
+    sa.Column('description', sa.Text(), nullable=True),
+    sa.Column('keyid', sa.String(length=16), nullable=True),
+    sa.Column('fingerprint', sa.String(length=40), nullable=True),
+    sa.Column('public_url', sa.Text(), nullable=True),
+    sa.Column('inserted', sa.DateTime(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('fingerprint'),
+    sa.UniqueConstraint('keyid')
+    )
+    op.create_table('sign_tasks',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('build_id', sa.Integer(), nullable=False),
+    sa.Column('sign_key_id', sa.Integer(), nullable=False),
+    sa.Column('status', sa.Integer(), nullable=True),
+    sa.Column('error_message', sa.Text(), nullable=True),
+    sa.Column('log_href', sa.Text(), nullable=True),
+    sa.ForeignKeyConstraint(['build_id'], ['builds.id'], ),
+    sa.ForeignKeyConstraint(['sign_key_id'], ['sign_keys.id'], ),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.add_column('builds', sa.Column('signed', sa.Boolean(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('builds', 'signed')
+    op.drop_table('sign_tasks')
+    op.drop_table('sign_keys')
+    # ### end Alembic commands ###
