+++ alws/routers/builds.py
@@ -8,7 +8,8 @@ from fastapi import (
     status,
 )
 
-from alws import database, crud
+from alws import database
+from alws.crud import build as build_crud, build_node
 from alws.dependencies import get_db, JWTBearer
 from alws.errors import DataNotFoundError
 from alws.schemas import build_schema
@@ -27,7 +28,8 @@ async def create_build(
             user: dict = Depends(JWTBearer()),
             db: database.Session = Depends(get_db)
         ):
-    db_build = await crud.create_build(db, build, user['identity']['user_id'])
+    db_build = await build_crud.create_build(
+        db, build, user['identity']['user_id'])
     return db_build
 
 
@@ -39,7 +41,7 @@ async def get_builds_per_page(
     db: database.Session = Depends(get_db),
 ):
     search_params = build_schema.BuildSearch(**request.query_params)
-    return await crud.get_builds(
+    return await build_crud.get_builds(
         db=db,
         page_number=pageNumber,
         search_params=search_params,
@@ -48,7 +50,7 @@ async def get_builds_per_page(
 
 @router.get('/{build_id}/', response_model=build_schema.Build)
 async def get_build(build_id: int, db: database.Session = Depends(get_db)):
-    db_build = await crud.get_builds(db, build_id)
+    db_build = await build_crud.get_builds(db, build_id)
     if db_build is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
@@ -60,13 +62,13 @@ async def get_build(build_id: int, db: database.Session = Depends(get_db)):
 @router.patch('/{build_id}/restart-failed', response_model=build_schema.Build)
 async def restart_failed_build_items(build_id: int,
                                      db: database.Session = Depends(get_db)):
-    return await crud.update_failed_build_items(db, build_id)
+    return await build_node.update_failed_build_items(db, build_id)
 
 
 @router.delete('/{build_id}/remove', status_code=204)
 async def remove_build(build_id: int, db: database.Session = Depends(get_db)):
     try:
-        result = await crud.remove_build_job(db, build_id)
+        result = await build_crud.remove_build_job(db, build_id)
     except DataNotFoundError:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
