+++ alws/routers/platforms.py
@@ -2,8 +2,9 @@ import typing
 
 from fastapi import APIRouter, Depends
 
+from alws import database
+from alws.crud import platform as pl_crud, repository
 from alws.dependencies import get_db, JWTBearer
-from alws import database, crud
 from alws.schemas import platform_schema
 
 
@@ -19,7 +20,7 @@ async def create_platform(
             platform: platform_schema.PlatformCreate,
             db: database.Session = Depends(get_db)
         ):
-    return await crud.create_platform(db, platform)
+    return await pl_crud.create_platform(db, platform)
 
 
 @router.put('/', response_model=platform_schema.Platform)
@@ -27,25 +28,27 @@ async def modify_platform(
             platform: platform_schema.PlatformModify,
             db: database.Session = Depends(get_db)
         ):
-    return await crud.modify_platform(db, platform)
+    return await pl_crud.modify_platform(db, platform)
 
 
 @router.get('/', response_model=typing.List[platform_schema.Platform])
 async def get_platforms(db: database.Session = Depends(get_db)):
-    return await crud.get_platforms(db)
+    return await pl_crud.get_platforms(db)
 
 
 @router.patch('/{platform_id}/add-repositories',
               response_model=platform_schema.Platform)
-async def add_repositories_to_platform(platform_id: int,
-                                       repositories_ids: typing.List[int],
-                                       db: database.Session = Depends(get_db)):
-    return await crud.add_to_platform(db, platform_id, repositories_ids)
+async def add_repositories_to_platform(
+        platform_id: int, repositories_ids: typing.List[int],
+        db: database.Session = Depends(get_db)):
+    return await repository.add_to_platform(
+        db, platform_id, repositories_ids)
 
 
 @router.patch('/{platform_id}/remove-repositories',
               response_model=platform_schema.Platform)
-async def remove_repositories_to_platform(platform_id: int,
-                                          repositories_ids: typing.List[int],
-                                          db: database.Session = Depends(get_db)):
-    return await crud.remove_from_platform(db, platform_id, repositories_ids)
+async def remove_repositories_to_platform(
+        platform_id: int, repositories_ids: typing.List[int],
+        db: database.Session = Depends(get_db)):
+    return await repository.remove_from_platform(
+        db, platform_id, repositories_ids)
