+++ alws/routers/releases.py
@@ -2,7 +2,8 @@ import typing
 
 from fastapi import APIRouter, Depends
 
-from alws import database, crud
+from alws import database
+from alws.crud import release as r_crud
 from alws.dependencies import get_db, JWTBearer
 from alws.schemas import release_schema
 
@@ -16,14 +17,14 @@ router = APIRouter(
 
 @router.get('/', response_model=typing.List[release_schema.Release])
 async def get_releases(db: database.Session = Depends(get_db)):
-    return await crud.get_releases(db)
+    return await r_crud.get_releases(db)
 
 
 @router.post('/new/', response_model=release_schema.Release)
 async def create_new_release(builds: release_schema.ReleaseCreate,
                              db: database.Session = Depends(get_db),
                              user: dict = Depends(JWTBearer())):
-    release = await crud.create_new_release(
+    release = await r_crud.create_new_release(
         db, user['identity']['user_id'], builds)
     return release
 
@@ -32,12 +33,12 @@ async def create_new_release(builds: release_schema.ReleaseCreate,
 async def update_release(release_id: int,
                          payload: release_schema.ReleaseUpdate,
                          db: database.Session = Depends(get_db)):
-    return await crud.update_release(db, release_id, payload)
+    return await r_crud.update_release(db, release_id, payload)
 
 
 @router.post('/{release_id}/commit/',
              response_model=release_schema.ReleaseCommitResult)
 async def commit_release(release_id: int,
                          db: database.Session = Depends(get_db)):
-    release, message = await crud.commit_release(db, release_id)
+    release, message = await r_crud.commit_release(db, release_id)
     return {'release': release, 'message': message}
