+++ alws/routers/repositories.py
@@ -2,7 +2,8 @@ import typing
 
 from fastapi import APIRouter, Depends
 
-from alws import database, crud
+from alws import database
+from alws.crud import repository
 from alws.dependencies import get_db, JWTBearer
 from alws.schemas import repository_schema
 
@@ -16,13 +17,13 @@ router = APIRouter(
 
 @router.get('/', response_model=typing.List[repository_schema.Repository])
 async def get_repositories(db: database.Session = Depends(get_db)):
-    return await crud.get_repositories(db)
+    return await repository.get_repositories(db)
 
 
 @router.get('/{repository_id}/',
             response_model=typing.Union[None, repository_schema.Repository])
 async def get_repository(repository_id: int, db: database.Session = Depends(get_db)):
-    result = await crud.get_repositories(db, repository_id=repository_id)
+    result = await repository.get_repositories(db, repository_id=repository_id)
     if result:
         return result[0]
     return None
